requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
click==8.1.7
feedparser==6.0.11
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
